add PhotoSchema

+++ src/controllers/image.ts
@@ -2,6 +2,9 @@ import { Request, Response } from "express";
 import path from "path";
 import fs from "fs";
 import { v4 as uuidv4 } from "uuid";
+import sharp from "sharp";
+import Photo from "../models/Photo";
+import { AdminAuthRequest } from "../middlewares/verify";
 
 const imageDirectory = (filename?: string) => {
   const basePath =
@@ -12,55 +15,100 @@ const imageDirectory = (filename?: string) => {
   return imagePath;
 };
 
-export const getImages = (req: Request, res: Response) => {
+export const getImages = async (req: Request, res: Response) => {
   try {
-    const imageFiles = fs.readdirSync(imageDirectory());
-    const images = imageFiles.map((fileName) => {
-      return {
-        filename: fileName,
-        url: `/public/images/${fileName}`,
-      };
+    const pageQueryParam = req.query.page;
+    const perPageQueryParam = req.query.perPage;
+
+    const page = pageQueryParam ? parseInt(pageQueryParam as string) : 1;
+    const perPage = perPageQueryParam
+      ? parseInt(perPageQueryParam as string)
+      : 40;
+
+    const totalPhotos = await Photo.countDocuments();
+    const totalPages = Math.ceil(totalPhotos / perPage);
+
+    const photos = await Photo.find()
+      .skip((page - 1) * perPage)
+      .limit(perPage);
+
+    const photoPromises = photos.map(async (i) => {
+      try {
+        return await i.toPhotoResponse();
+      } catch (error) {
+        console.error("Error processing photo:", error);
+        return null;
+      }
     });
 
-    res.status(200).json(images);
+    const photo = await Promise.all(photoPromises);
+
+    const nextPage = page < totalPages ? page + 1 : null;
+    const prevPage = page > 1 ? page - 1 : null;
+
+    res.status(200).json({
+      photo,
+      totalPages,
+      totalPhotos,
+      currentPage: page,
+      nextPage,
+      prevPage,
+    });
   } catch (error) {
     console.error("Error retrieving images:", error);
     res.status(500).send("Error retrieving images.");
   }
 };
 
-export const saveImage = (req: Request, res: Response) => {
+export const saveImage = async (req: Request, res: Response) => {
   try {
     const filename = req.header("filename");
     if (!filename) {
       return res.status(400).send("No filename specified");
     }
 
+    const newReq = req as unknown as AdminAuthRequest;
+
     const uniqueSuffix = uuidv4();
     const sanitizedOriginalName = filename.replace(/[^a-zA-Z0-9]/g, "_");
     const uniqueFilename = `${sanitizedOriginalName}-${uniqueSuffix}.jpg`;
 
-    fs.mkdirSync(imageDirectory(), { recursive: true });
     fs.writeFileSync(imageDirectory(uniqueFilename), req.body);
-
-    res.status(200).json({ message: "Image uploaded and saved successfully." });
+    const imageInfo = await sharp(
+      `${imageDirectory()}/${uniqueFilename}`
+    ).metadata();
+    const photo = await Photo.create({
+      url: `/public/images/${uniqueFilename}`,
+      width: imageInfo.width,
+      height: imageInfo.height,
+      filename: uniqueFilename,
+      alt: req.body.alt,
+      uploader: newReq.userId,
+    });
+    res.status(200).json({ photo: await photo.toPhotoResponse() });
   } catch (error) {
     console.error("Error saving image:", error);
     res.status(500).send("Error saving image.");
   }
 };
 
-export const deleteImage = (req: Request, res: Response) => {
+export const deleteImage = async (req: Request, res: Response) => {
   try {
     const filename = req.params.filename;
     if (!filename) {
       return res.status(400).json({ message: "No filename specified" });
     }
 
+    const deletePhoto = await Photo.findOneAndDelete({ filename });
+
+    if (!deletePhoto) {
+      return res.status(404).json({ message: "photo not found" });
+    }
+
     if (fs.existsSync(imageDirectory(filename))) {
       fs.unlinkSync(imageDirectory(filename));
 
-      res.status(200).json({ message: "Image deleted successfully." });
+      res.status(204).json({ message: "Image deleted successfully." });
     } else {
       res.status(404).json({ message: "Image not found" });
     }
